refactor: table types
