+++ projects/ion/src/lib/table/table.component.ts
@@ -1,17 +1,9 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { CheckBoxStates } from '../core/types/checkbox';
 import { PageEvent } from '../core/types/pagination';
+import { TableEvent } from '../core/types/table';
 import { SafeAny } from '../utils/safe-any';
 import { ActionTable, Column, ConfigTable, TableUtils } from './utilsTable';
-
-interface TableEvent {
-  rows_selected: SafeAny[];
-}
-
-export interface IonTableProps<T> {
-  config: ConfigTable<T>;
-  events?: EventEmitter<TableEvent>;
-}
 
 const stateChange = {
   checked: 'enabled',
